requirements.txt
numpy
pillow
matplotlib
tqdm
omnifig
webcolors
scikit-learn
git+git://github.com/aparkins/pydip